verity-ai-backend/requirements.txt
pandas
numpy
scikit-learn
supabase
python-dotenv
requests
fastapi
uvicorn